Move getting current time function from SpeakerFragment to ListenerFragment
