+++ app/src/main/res/layout/fragment_speaker.xml
@@ -25,25 +25,6 @@
         android:textSize="25dp"
         android:textStyle="bold" />
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/textViewTime_Speaker"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="20dp"
-            android:text="Current time : "
-            android:textSize="20dp" />
-
-        <TextView
-            android:id="@+id/textViewTimeEdit_Speaker"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            android:textSize="20dp" />
-    </LinearLayout>
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -90,7 +71,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:layout_margin="60dp"
+        android:layout_margin="100dp"
         android:background="@drawable/btn_background"
         android:paddingHorizontal="30dp"
         android:text="CHECK IN"
